Fix cLib_calcTimer (#1963)
* Fix cLib_calcTimer

* Clean up some additional RELs

+++ include/SSystem/SComponent/c_lib.h
@@ -87,7 +87,8 @@ inline T cLib_maxLimit(T val, T max) {
 
 template <typename T>
 T cLib_calcTimer(T* value) {
-    if (*value != 0) {
+    // Casting 0 to u16 may not be correct, but is matching for now
+    if (*value != (u16)0) {
         *value = *value - 1;
     }
     return *value;
